+++ src/main/java/ru/itpark/Main.java
@@ -3,7 +3,7 @@ package ru.itpark;
 public class Main {
     public static void main(String[] args) {
 
-        double sumDeposit = (double)100_000;
+        int sumDeposit = 100_000;
         int period = 14;
         double percentDeposit = 10;
         String dateStart = "21.10.2018";
